fix: handle reviews
