+++ web-marketplace/src/pages/Dashboard/MyCreditBatches/MyCreditBatches.tsx
@@ -6,15 +6,16 @@ import OutlinedButton from 'web-components/lib/components/buttons/OutlinedButton
 import EmptyState from 'web-components/lib/components/empty-state';
 import NoEcocreditsIcon from 'web-components/lib/components/icons/NoEcocreditsIcon';
 import PlusIcon from 'web-components/lib/components/icons/PlusIcon';
-import { Label } from 'web-components/lib/components/typography';
 
 import { useWallet } from 'lib/wallet/wallet';
 
 import WithLoader from 'components/atoms/WithLoader';
 import { CreditBatches } from 'components/organisms';
 import { useFetchPaginatedBatches } from 'hooks/batches/useFetchPaginatedBatches';
+import { usePaginatedBatchesByIssuer } from 'hooks/batches/usePaginatedBatchesByIssuer';
 
 import { NO_CREDIT_BATCHES_MESSAGE } from './MyCreditBatches.constants';
+import { MyCreditBatchesTable } from './MyCreditBatches.Table';
 
 export const MyCreditBatches = (): JSX.Element => {
   const theme = useTheme();
@@ -25,49 +26,12 @@ export const MyCreditBatches = (): JSX.Element => {
 
   return (
     <Box sx={{ width: '100%' }}>
-      <WithLoader
-        isLoading={batchesWithSupply === undefined}
-        sx={{
-          display: 'flex',
-          justifyContent: 'center',
-          flexDirection: 'col',
-          alignItems: 'center',
-        }}
-      >
-        <>
-          <Box
-            sx={{
-              display: hasNoBatches ? 'none' : 'flex',
-              flexDirection: ['column', 'row'],
-              justifyContent: 'space-between',
-              alignItems: ['flex-start', 'center'],
-              mb: 10.5,
-            }}
-          >
-            <Label
-              sx={{
-                color: 'info.dark',
-                mb: [3.5, 0],
-              }}
-            >
-              Successfully issued credit batches
-            </Label>
-            <OutlinedButton
-              startIcon={<PlusIcon color={theme.palette.secondary.main} />}
-              component={Link}
-              to="/profile/create-batch"
-            >
-              create credit batch
-            </OutlinedButton>
-          </Box>
-          <CreditBatches
-            creditBatches={batchesWithSupply}
-            onTableChange={setPaginationParams}
-            initialPaginationParams={paginationParams}
-            isIgnoreOffset
-          />
-        </>
-      </WithLoader>
+      <MyCreditBatchesTable
+        hasNoBatches={hasNoBatches}
+        setPaginationParams={setPaginationParams}
+        batchesWithSupply={batchesWithSupply}
+        useCreate
+      />
       {hasNoBatches && (
         <EmptyState
           message={NO_CREDIT_BATCHES_MESSAGE}
