+++ web-marketplace/src/pages/EcocreditsByAccount/EcocreditsByAccount.tsx
@@ -1,9 +1,11 @@
-import { Box } from '@mui/material';
 import { useMemo } from 'react';
 import { Outlet, useLocation, useParams } from 'react-router-dom';
+import { Box } from '@mui/material';
 
 import { Flex } from 'web-components/lib/components/box';
 import BridgeIcon from 'web-components/lib/components/icons/BridgeIcon';
+import { CreditBatchIcon } from 'web-components/lib/components/icons/CreditBatchIcon';
+import { CreditClassIcon } from 'web-components/lib/components/icons/CreditClassIcon';
 import CreditsIcon from 'web-components/lib/components/icons/CreditsIcon';
 import { ProjectPageIcon } from 'web-components/lib/components/icons/ProjectPageIcon';
 import { ProfileHeader } from 'web-components/lib/components/organisms/ProfileHeader/ProfileHeader';
@@ -13,10 +15,9 @@ import { IconTabs } from 'web-components/lib/components/tabs/IconTabs';
 import { truncate } from 'web-components/lib/utils/truncate';
 
 import { getAccountUrl } from 'lib/block-explorer';
+import { isBridgeEnabled } from 'lib/ledger';
 import { getProfileLink } from 'lib/profileLink';
 
-import { Link } from 'components/atoms';
-import WithLoader from 'components/atoms/WithLoader';
 import {
   getSocialsLinks,
   getUserImages,
@@ -25,15 +26,14 @@ import {
   DEFAULT_NAME,
   profileVariantMapping,
 } from 'pages/ProfileEdit/ProfileEdit.constants';
-
+import { Link } from 'components/atoms';
+import WithLoader from 'components/atoms/WithLoader';
 import { useQueryIsIssuer } from 'hooks/useQueryIsIssuer';
 import { useQueryIsProjectAdmin } from 'hooks/useQueryIsProjectAdmin';
-import { CreditBatchIcon } from 'web-components/lib/components/icons/CreditBatchIcon';
-import { CreditClassIcon } from 'web-components/lib/components/icons/CreditClassIcon';
+
 import { ProfileNotFound } from './EcocreditsByAccount.NotFound';
 import { ecocreditsByAccountStyles } from './EcocreditsByAccount.styles';
 import { useProfileData } from './hooks/useProfileData';
-import { isBridgeEnabled } from 'lib/ledger';
 
 export const EcocreditsByAccount = (): JSX.Element => {
   const { accountAddressOrId } = useParams<{ accountAddressOrId: string }>();
@@ -85,7 +85,7 @@ export const EcocreditsByAccount = (): JSX.Element => {
         hidden: !isBridgeEnabled,
       },
     ],
-    [accountAddressOrId, party, isIssuer, isProjectAdmin],
+    [accountAddressOrId, isIssuer, isProjectAdmin],
   );
 
   const activeTab = Math.max(
